+++ android/versioned-abis/expoview-abi49_0_0/src/main/java/abi49_0_0/expo/modules/av/video/FullscreenVideoPlayerPresentationChangeListener.java
@@ -1,8 +1,0 @@
-package abi49_0_0.expo.modules.av.video;
-
-public interface FullscreenVideoPlayerPresentationChangeListener {
-  void onFullscreenPlayerWillPresent();
-  void onFullscreenPlayerDidPresent();
-  void onFullscreenPlayerWillDismiss();
-  void onFullscreenPlayerDidDismiss();
-}