+++ android/versioned-abis/expoview-abi49_0_0/src/main/java/abi49_0_0/expo/modules/brightness/Exceptions.kt
@@ -1,6 +1,0 @@
-package abi49_0_0.expo.modules.brightness
-
-import abi49_0_0.expo.modules.kotlin.exception.CodedException
-
-class BrightnessPermissionsException :
-  CodedException("WRITE_SETTINGS permission has not been granted")