+++ android/versioned-abis/expoview-abi49_0_0/src/main/java/abi49_0_0/expo/modules/image/Exceptions.kt
@@ -1,9 +1,0 @@
-package abi49_0_0.expo.modules.image
-
-import abi49_0_0.expo.modules.kotlin.exception.CodedException
-
-class ImagePrefetchFailure(message: String) :
-  CodedException(message = "Failed to prefetch the image: $message")
-
-class MissingActivity :
-  CodedException(message = "The current activity is no longer available")