+++ android/versioned-abis/expoview-abi49_0_0/src/main/java/abi49_0_0/expo/modules/imagemanipulator/actions/Action.kt
@@ -1,7 +1,0 @@
-package abi49_0_0.expo.modules.imagemanipulator.actions
-
-import android.graphics.Bitmap
-
-interface Action {
-  fun run(bitmap: Bitmap): Bitmap
-}