+++ android/versioned-abis/expoview-abi49_0_0/src/main/java/abi49_0_0/expo/modules/kotlin/records/Record.kt
@@ -1,3 +1,0 @@
-package abi49_0_0.expo.modules.kotlin.records
-
-interface Record