+++ android/versioned-abis/expoview-abi49_0_0/src/main/java/abi49_0_0/expo/modules/kotlin/views/ViewWrapperDelegateHolder.kt
@@ -1,5 +1,0 @@
-package abi49_0_0.expo.modules.kotlin.views
-
-interface ViewWrapperDelegateHolder {
-  val viewWrapperDelegate: ViewManagerWrapperDelegate
-}