+++ ios/vendored/sdk49/@stripe/stripe-react-native/ios/StripeContainerManager.m
@@ -1,8 +1,0 @@
-#import <Foundation/Foundation.h>
-#import <ABI49_0_0React/ABI49_0_0RCTBridgeModule.h>
-#import <ABI49_0_0React/ABI49_0_0RCTViewManager.h>
-#import "ABI49_0_0React/ABI49_0_0RCTUIManager.h"
-
-@interface ABI49_0_0RCT_EXTERN_REMAP_MODULE(StripeContainerManager, ABI49_0_0StripeContainerManager, ABI49_0_0RCTViewManager)
-ABI49_0_0RCT_EXPORT_VIEW_PROPERTY(keyboardShouldPersistTaps, BOOL)
-@end