+++ ios/vendored/sdk49/react-native-gesture-handler/ios/Handlers/ABI49_0_0RNManualHandler.h
@@ -1,4 +1,0 @@
-#import "ABI49_0_0RNGestureHandler.h"
-
-@interface ABI49_0_0RNManualGestureHandler : ABI49_0_0RNGestureHandler
-@end