+++ ios/vendored/sdk49/react-native-safe-area-context/ios/ABI49_0_0RNCSafeAreaViewManager.h
@@ -1,9 +1,0 @@
-#import <ABI49_0_0React/ABI49_0_0RCTViewManager.h>
-
-NS_ASSUME_NONNULL_BEGIN
-
-@interface ABI49_0_0RNCSafeAreaViewManager : ABI49_0_0RCTViewManager
-
-@end
-
-NS_ASSUME_NONNULL_END