+++ ios/versioned/sdk49/EXFont/EXFont/Singletons/ABI49_0_0EXFontScalersManager.h
@@ -1,8 +1,0 @@
-// Copyright 2018-present 650 Industries. All rights reserved.
-
-#import <ABI49_0_0ExpoModulesCore/ABI49_0_0EXSingletonModule.h>
-#import <ABI49_0_0ExpoModulesCore/ABI49_0_0EXFontScalersManagerInterface.h>
-
-@interface ABI49_0_0EXFontScalersManager : ABI49_0_0EXSingletonModule <ABI49_0_0EXFontScalersManagerInterface>
-
-@end