+++ ios/versioned/sdk49/EXSensors/EXSensors/Modules/SensorModules/ABI49_0_0EXDeviceMotion.h
@@ -1,7 +1,0 @@
-// Copyright 2015-present 650 Industries. All rights reserved.
-
-#import <ABI49_0_0EXSensors/ABI49_0_0EXBaseSensorModule.h>
-
-@interface ABI49_0_0EXDeviceMotion : ABI49_0_0EXBaseSensorModule
-
-@end