+++ ios/versioned/sdk49/EXUpdates/EXUpdates/ABI49_0_0EXUpdates.h
@@ -1,5 +1,0 @@
-// Copyright 2018-present 650 Industries. All rights reserved.
-
-#if __has_include("ABI49_0_0EXUpdates-umbrella.h")
-#import "ABI49_0_0EXUpdates-umbrella.h"
-#endif